Adjust layout of plugin code to hide implementation files to make importing the plugin a lot easier. Renamed classes that had the platform as part of the name to use a prefix convention instead of a suffix one
