+++ CHANGELOG.md
@@ -10,7 +10,6 @@
 
 * Fix changelog
 
-
 ## [0.0.4]
 
 * Updated README
@@ -51,59 +50,88 @@
 * [Android] Bug fix in calculating when to show a scheduled  notification. Ensure scheduled Android notifications will remain scheduled even after rebooting.
 
 ## [0.1.3]
+
 * Fix broken example app for iOS due to incorrect reference to custom sound file. Added ability to handle when a notification is tapped. See updated example for details on how to do this and will navigate to another page. Note that the second page isn't rendering full-screen on Android if the notification was tapped on while the app was in the foreground. Suspect that this is Flutter rendering issue and have logged this on the Flutter repository at https://github.com/flutter/flutter/issues/16636
 
 ## [0.1.4]
+
 * [Android] Add inbox notification style
 
 ## [0.1.5]
+
 * **BREAKING CHANGES** There are no functional changes. This is an API cleanup release where I've reorganised the Dart classes to better separate them by platform. What this means is that the import statements in your coode will need to be fixed. Apologies to anyone using the plugin but I feel that this was necessary as Flutter may target additional platforms in the future. Hopefully you'll agree that the end result looks a better :)
 
 ## [0.1.6]
+
 * **BREAKING CHANGES** Apologies again, this is another cleanup release. FlutterLocalNotifications class has been renamed to FlutterLocalNotificationsPlugin now as it makes more sense from a readability perspective. More importantly, the class and methods are also no longer static to faciliate mocking and testing. It's something I should've picked up on earlier so sorry once again. Check the source code for an example on how to mock the plugin when testing
 
 ## [0.1.7]
+
 * [Android] Add ability to show grouped notifications. Example code has been updated to demonstrate this functionality.
 * Fixed the example project so it works with the new release of Cocoapods (1.5.0)
 * Fixes for when API methods were called without specifying platform specific settings
 
 ## [0.1.8]
+
 * [Android] Bug fix for grouping notifications
 
 ## [0.1.9]
+
 * [Android/iOS] Add ability to cancel/remove all notifications
 
 ## [0.2.0]
+
 * [Android] Add ability to specify if notifications should automatically be dismissed upon touching them
 * [Android] Add ability to specify in notifications are ongoing
 * [Android] Fix bug in cancelling all notifications
 
 ## [0.2.1]
+
 * [Android/iOS] Add ability to set a notification to be periodically displayed
 * [Android] Fix a bug where the small icon could not be be found when loading scheduled notifications from shared preferences to reschedule them on a device reboot
 * [Android] Fix example app manifest file
 
 ## [0.2.2]
+
 * [Android/iOS] Fix RepeatInterval not being mapped to the correct values on the native side. Thanks to Thibault Deckers (deckerst) for spotting the issue.
 
 ## [0.2.3]
+
 * [Android/iOS] Add ability to have a notification shown daily at a specified time. Credits to Javier Lecuona (javiercbk) for submitting the PR for this.
 * [Android/iOS] Add ability to have a notification shown weekly on a specific day and time.
 
 ## [0.2.4]
+
 * [Android] Add ability to set the colour.
 
 ## [0.2.5]
+
 * [Android] Bug fix for previous release
 
 ## [0.2.6]
+
 * [Android] Fix bug with applying ongoing and autoCancel properties
 
 ## [0.2.7]
+
 * [Android] Fix issue with showDailyAtTime and showWeeklyAtDayAndTime where time occurred in the past and caused notification to trigger instantly
 
 ## [0.2.8]
+
 * Update dev dependencies required for flutter test to run. This is due to a breaking change in Flutter. See https://github.com/flutter/flutter/wiki/Changelog
 
 ## [0.2.9]
+
 * Fix error in calling initialize error on iOS versions < 9
+
+## [0.3.0]
+
+* **BREAKING CHANGES** restructured code so that only a single import statement is now needed and classes that had the platform (Android/iOS) as a suffix are now prefixes to improve readability of code and follow the recommendations for writing Dart code i.e. write code that reads more like a sentence. The following have been renamed
+
+    * InitializationSettingsAndroid -> AndroidInitializationSettings
+    * InitializationSettingsIOS -> IOSInitializationSettings
+    * NotificationDetailsAndroid -> AndroidNotificationDetails
+    * NotificationStyleAndroid -> AndroidNotificationStyle
+    * NotificationDetailsIOS -> IOSNotificationDetails
+
+
