+++ example/lib/main.dart
@@ -2,15 +2,6 @@ import 'dart:async';
 import 'dart:typed_data';
 import 'package:flutter/material.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
-import 'package:flutter_local_notifications/initialization_settings.dart';
-import 'package:flutter_local_notifications/notification_details.dart';
-import 'package:flutter_local_notifications/platform_specifics/android/initialization_settings_android.dart';
-import 'package:flutter_local_notifications/platform_specifics/android/notification_details_android.dart';
-import 'package:flutter_local_notifications/platform_specifics/android/styles/big_text_style_information.dart';
-import 'package:flutter_local_notifications/platform_specifics/android/styles/default_style_information.dart';
-import 'package:flutter_local_notifications/platform_specifics/android/styles/inbox_style_information.dart';
-import 'package:flutter_local_notifications/platform_specifics/ios/initialization_settings_ios.dart';
-import 'package:flutter_local_notifications/platform_specifics/ios/notification_details_ios.dart';
 
 void main() {
   runApp(
@@ -30,8 +21,8 @@ class _MyAppState extends State<MyApp> {
     super.initState();
     // initialise the plugin
     var initializationSettingsAndroid =
-        new InitializationSettingsAndroid('app_icon');
-    var initializationSettingsIOS = new InitializationSettingsIOS();
+        new AndroidInitializationSettings('app_icon');
+    var initializationSettingsIOS = new IOSInitializationSettings();
     var initializationSettings = new InitializationSettings(
         initializationSettingsAndroid, initializationSettingsIOS);
     flutterLocalNotificationsPlugin = new FlutterLocalNotificationsPlugin();
@@ -152,10 +143,10 @@ class _MyAppState extends State<MyApp> {
   }
 
   Future _showNotification() async {
-    var androidPlatformChannelSpecifics = new NotificationDetailsAndroid(
+    var androidPlatformChannelSpecifics = new AndroidNotificationDetails(
         'your channel id', 'your channel name', 'your channel description',
         importance: Importance.Max, priority: Priority.High);
-    var iOSPlatformChannelSpecifics = new NotificationDetailsIOS();
+    var iOSPlatformChannelSpecifics = new IOSNotificationDetails();
     var platformChannelSpecifics = new NotificationDetails(
         androidPlatformChannelSpecifics, iOSPlatformChannelSpecifics);
     await flutterLocalNotificationsPlugin.show(
@@ -177,7 +168,7 @@ class _MyAppState extends State<MyApp> {
     vibrationPattern[2] = 5000;
     vibrationPattern[3] = 2000;
 
-    var androidPlatformChannelSpecifics = new NotificationDetailsAndroid(
+    var androidPlatformChannelSpecifics = new AndroidNotificationDetails(
         'your other channel id',
         'your other channel name',
         'your other channel description',
@@ -186,7 +177,7 @@ class _MyAppState extends State<MyApp> {
         vibrationPattern: vibrationPattern,
         color: const Color.fromARGB(255, 255, 0, 0));
     var iOSPlatformChannelSpecifics =
-        new NotificationDetailsIOS(sound: "slow_spring_board.aiff");
+        new IOSNotificationDetails(sound: "slow_spring_board.aiff");
     var platformChannelSpecifics = new NotificationDetails(
         androidPlatformChannelSpecifics, iOSPlatformChannelSpecifics);
     await flutterLocalNotificationsPlugin.schedule(
@@ -198,14 +189,14 @@ class _MyAppState extends State<MyApp> {
   }
 
   Future _showNotificationWithNoSound() async {
-    var androidPlatformChannelSpecifics = new NotificationDetailsAndroid(
+    var androidPlatformChannelSpecifics = new AndroidNotificationDetails(
         'silent channel id',
         'silent channel name',
         'silent channel description',
         playSound: false,
         styleInformation: new DefaultStyleInformation(true, true));
     var iOSPlatformChannelSpecifics =
-        new NotificationDetailsIOS(presentSound: false);
+        new IOSNotificationDetails(presentSound: false);
     var platformChannelSpecifics = new NotificationDetails(
         androidPlatformChannelSpecifics, iOSPlatformChannelSpecifics);
     await flutterLocalNotificationsPlugin.show(0, '<b>silent</b> title',
@@ -220,11 +211,11 @@ class _MyAppState extends State<MyApp> {
         htmlFormatContentTitle: true,
         summaryText: 'summary <i>text</i>',
         htmlFormatSummaryText: true);
-    var androidPlatformChannelSpecifics = new NotificationDetailsAndroid(
+    var androidPlatformChannelSpecifics = new AndroidNotificationDetails(
         'big text channel id',
         'big text channel name',
         'big text channel description',
-        style: NotificationStyleAndroid.BigText,
+        style: AndroidNotificationStyle.BigText,
         styleInformation: bigTextStyleInformation);
     var platformChannelSpecifics =
         new NotificationDetails(androidPlatformChannelSpecifics, null);
@@ -242,9 +233,9 @@ class _MyAppState extends State<MyApp> {
         htmlFormatContentTitle: true,
         summaryText: 'summary <i>text</i>',
         htmlFormatSummaryText: true);
-    var androidPlatformChannelSpecifics = new NotificationDetailsAndroid(
+    var androidPlatformChannelSpecifics = new AndroidNotificationDetails(
         'inbox channel id', 'inboxchannel name', 'inbox channel description',
-        style: NotificationStyleAndroid.Inbox,
+        style: AndroidNotificationStyle.Inbox,
         styleInformation: inboxStyleInformation);
     var platformChannelSpecifics =
         new NotificationDetails(androidPlatformChannelSpecifics, null);
@@ -258,7 +249,7 @@ class _MyAppState extends State<MyApp> {
     var groupChannelName = 'grouped channel name';
     var groupChannelDescription = 'grouped channel description';
     // example based on https://developer.android.com/training/notify-user/group.html
-    var firstNotificationAndroidSpecifics = new NotificationDetailsAndroid(
+    var firstNotificationAndroidSpecifics = new AndroidNotificationDetails(
         groupChannelId, groupChannelName, groupChannelDescription,
         importance: Importance.Max,
         priority: Priority.High,
@@ -267,7 +258,7 @@ class _MyAppState extends State<MyApp> {
         new NotificationDetails(firstNotificationAndroidSpecifics, null);
     await flutterLocalNotificationsPlugin.show(1, 'Alex Faarborg',
         'You will not believe...', firstNotificationPlatformSpecifics);
-    var secondNotificationAndroidSpecifics = new NotificationDetailsAndroid(
+    var secondNotificationAndroidSpecifics = new AndroidNotificationDetails(
         groupChannelId, groupChannelName, groupChannelDescription,
         importance: Importance.Max,
         priority: Priority.High,
@@ -286,9 +277,9 @@ class _MyAppState extends State<MyApp> {
     lines.add('Jeff Chang    Launch Party');
     var inboxStyleInformation = new InboxStyleInformation(lines,
         contentTitle: '2 new messages', summaryText: 'janedoe@example.com');
-    var androidPlatformChannelSpecifics = new NotificationDetailsAndroid(
+    var androidPlatformChannelSpecifics = new AndroidNotificationDetails(
         groupChannelId, groupChannelName, groupChannelDescription,
-        style: NotificationStyleAndroid.Inbox,
+        style: AndroidNotificationStyle.Inbox,
         styleInformation: inboxStyleInformation,
         groupKey: groupKey,
         setAsGroupSummary: true);
@@ -314,13 +305,13 @@ class _MyAppState extends State<MyApp> {
   }
 
   Future _showOngoingNotification() async {
-    var androidPlatformChannelSpecifics = new NotificationDetailsAndroid(
+    var androidPlatformChannelSpecifics = new AndroidNotificationDetails(
         'your channel id', 'your channel name', 'your channel description',
         importance: Importance.Max,
         priority: Priority.High,
         ongoing: true,
         autoCancel: false);
-    var iOSPlatformChannelSpecifics = new NotificationDetailsIOS();
+    var iOSPlatformChannelSpecifics = new IOSNotificationDetails();
     var platformChannelSpecifics = new NotificationDetails(
         androidPlatformChannelSpecifics, iOSPlatformChannelSpecifics);
     await flutterLocalNotificationsPlugin.show(0, 'ongoing notification title',
@@ -328,11 +319,11 @@ class _MyAppState extends State<MyApp> {
   }
 
   Future _repeatNotification() async {
-    var androidPlatformChannelSpecifics = new NotificationDetailsAndroid(
+    var androidPlatformChannelSpecifics = new AndroidNotificationDetails(
         'repeating channel id',
         'repeating channel name',
         'repeating description');
-    var iOSPlatformChannelSpecifics = new NotificationDetailsIOS();
+    var iOSPlatformChannelSpecifics = new IOSNotificationDetails();
     var platformChannelSpecifics = new NotificationDetails(
         androidPlatformChannelSpecifics, iOSPlatformChannelSpecifics);
     await flutterLocalNotificationsPlugin.periodicallyShow(0, 'repeating title',
@@ -341,11 +332,11 @@ class _MyAppState extends State<MyApp> {
 
   Future _showDailyAtTime() async {
     var time = new Time(10, 0, 0);
-    var androidPlatformChannelSpecifics = new NotificationDetailsAndroid(
+    var androidPlatformChannelSpecifics = new AndroidNotificationDetails(
         'repeatDailyAtTime channel id',
         'repeatDailyAtTime channel name',
         'repeatDailyAtTime description');
-    var iOSPlatformChannelSpecifics = new NotificationDetailsIOS();
+    var iOSPlatformChannelSpecifics = new IOSNotificationDetails();
     var platformChannelSpecifics = new NotificationDetails(
         androidPlatformChannelSpecifics, iOSPlatformChannelSpecifics);
     await flutterLocalNotificationsPlugin.showDailyAtTime(
@@ -358,11 +349,11 @@ class _MyAppState extends State<MyApp> {
 
   Future _showWeeklyAtDayAndTime() async {
     var time = new Time(10, 0, 0);
-    var androidPlatformChannelSpecifics = new NotificationDetailsAndroid(
+    var androidPlatformChannelSpecifics = new AndroidNotificationDetails(
         'show weekly channel id',
         'show weekly channel name',
         'show weekly description');
-    var iOSPlatformChannelSpecifics = new NotificationDetailsIOS();
+    var iOSPlatformChannelSpecifics = new IOSNotificationDetails();
     var platformChannelSpecifics = new NotificationDetails(
         androidPlatformChannelSpecifics, iOSPlatformChannelSpecifics);
     await flutterLocalNotificationsPlugin.showWeeklyAtDayAndTime(
