+++ lib/src/initialization_settings.dart
@@ -1,0 +1,12 @@
+part of flutter_local_notifications;
+
+/// Settings for initializing the plugin for each platform
+class InitializationSettings {
+  /// Settings for Android
+  final AndroidInitializationSettings android;
+
+  /// Settings for iOS
+  final IOSInitializationSettings ios;
+
+  const InitializationSettings(this.android, this.ios);
+}
