+++ lib/src/platform_specifics/android/initialization_settings.dart
@@ -1,9 +1,11 @@
+part of flutter_local_notifications;
+
 /// Plugin initialization settings for Android
-class InitializationSettingsAndroid {
+class AndroidInitializationSettings {
   /// Sets the default icon for notifications
   final String defaultIcon;
 
-  const InitializationSettingsAndroid(this.defaultIcon);
+  const AndroidInitializationSettings(this.defaultIcon);
 
   Map<String, dynamic> toMap() {
     return <String, dynamic>{'defaultIcon': this.defaultIcon};
