+++ lib/src/platform_specifics/android/styles/style_information.dart
@@ -1,3 +1,5 @@
+part of flutter_local_notifications;
+
 abstract class StyleInformation {
   StyleInformation();
