+++ test/flutter_local_notifications_test.dart
@@ -3,12 +3,6 @@ import 'package:mockito/mockito.dart';
 import 'package:platform/platform.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
-import 'package:flutter_local_notifications/initialization_settings.dart';
-import 'package:flutter_local_notifications/notification_details.dart';
-import 'package:flutter_local_notifications/platform_specifics/android/initialization_settings_android.dart';
-import 'package:flutter_local_notifications/platform_specifics/android/notification_details_android.dart';
-import 'package:flutter_local_notifications/platform_specifics/ios/initialization_settings_ios.dart';
-import 'package:flutter_local_notifications/platform_specifics/ios/notification_details_ios.dart';
 
 void main() {
   MockMethodChannel mockChannel;
@@ -26,8 +20,8 @@ void main() {
               mockChannel, new FakePlatform(operatingSystem: 'ios'));
     });
     test('initialise plugin on iOS', () async {
-      const InitializationSettingsIOS initializationSettingsIOS =
-          InitializationSettingsIOS();
+      const IOSInitializationSettings initializationSettingsIOS =
+          IOSInitializationSettings();
       const InitializationSettings initializationSettings =
           InitializationSettings(null, initializationSettingsIOS);
       await flutterLocalNotificationsPlugin.initialize(initializationSettings);
@@ -35,8 +29,8 @@ void main() {
           'initialize', initializationSettingsIOS.toMap()));
     });
     test('show notification on iOS', () async {
-      NotificationDetailsIOS iOSPlatformChannelSpecifics =
-          new NotificationDetailsIOS();
+      IOSNotificationDetails iOSPlatformChannelSpecifics =
+          new IOSNotificationDetails();
       NotificationDetails platformChannelSpecifics =
           new NotificationDetails(null, iOSPlatformChannelSpecifics);
 
@@ -53,8 +47,8 @@ void main() {
     test('schedule notification on iOS', () async {
       var scheduledNotificationDateTime =
           new DateTime.now().add(new Duration(seconds: 5));
-      NotificationDetailsIOS iOSPlatformChannelSpecifics =
-          new NotificationDetailsIOS();
+      IOSNotificationDetails iOSPlatformChannelSpecifics =
+          new IOSNotificationDetails();
       NotificationDetails platformChannelSpecifics =
           new NotificationDetails(null, iOSPlatformChannelSpecifics);
       await flutterLocalNotificationsPlugin.schedule(id, title, body,
@@ -84,8 +78,8 @@ void main() {
               mockChannel, new FakePlatform(operatingSystem: 'android'));
     });
     test('initialise plugin on Android', () async {
-      const InitializationSettingsAndroid initializationSettingsAndroid =
-          InitializationSettingsAndroid('app_icon');
+      const AndroidInitializationSettings initializationSettingsAndroid =
+          AndroidInitializationSettings('app_icon');
       const InitializationSettings initializationSettings =
           InitializationSettings(initializationSettingsAndroid, null);
       await flutterLocalNotificationsPlugin.initialize(initializationSettings);
@@ -93,8 +87,8 @@ void main() {
           'initialize', initializationSettingsAndroid.toMap()));
     });
     test('show notification on Android', () async {
-      NotificationDetailsAndroid androidPlatformChannelSpecifics =
-          new NotificationDetailsAndroid('your channel id', 'your channel name',
+      AndroidNotificationDetails androidPlatformChannelSpecifics =
+          new AndroidNotificationDetails('your channel id', 'your channel name',
               'your channel description',
               importance: Importance.Max, priority: Priority.High);
 
@@ -115,8 +109,8 @@ void main() {
       var scheduledNotificationDateTime =
           new DateTime.now().add(new Duration(seconds: 5));
 
-      NotificationDetailsAndroid androidPlatformChannelSpecifics =
-          new NotificationDetailsAndroid('your other channel id',
+      AndroidNotificationDetails androidPlatformChannelSpecifics =
+          new AndroidNotificationDetails('your other channel id',
               'your other channel name', 'your other channel description');
 
       NotificationDetails platformChannelSpecifics =
